pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "specter-static-site"
version = "1.0.0"
description = "Reusable CDK construct for static site hosting on AWS"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib~=2.208.0",
    "constructs~=10.4.2",
    "cdk-nag~=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest~=8.0.0",
]
